Server/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
motor==3.1.1
pymongo==4.5.0
email-validator==2.1.0
aiohttp==3.9.1
pydantic==2.5.2
pydantic-settings==2.1.0
resend==0.6.0
fastapi-mail==1.4.1
aiofiles==23.2.1
# vapi-sdk==0.1.5
tensorflow-cpu
joblib
numpy